requirements-base.txt
# 基础依赖包 - 核心运行时依赖
# Web框架
Flask>=2.3.0
Flask-RESTful>=0.3.10
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0

# WebSocket和异步支持
eventlet>=0.33.0

# 数据库相关
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
cryptography>=46.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0

# 数据处理核心
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# 配置管理
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 工具库
requests>=2.31.0
python-dateutil>=2.8.0
pypinyin>=0.53.0


# 时序数据库客户端
influxdb-client>=1.38.0
influxdb>=5.3.1